// frontend/src/types/enums.ts
export enum SectionType {
  SINGLE_PROPERTY_AD = 'SINGLE_PROPERTY_AD',
  FEATURED_PROPERTY_AD = 'FEATURED_PROPERTY_AD',
  MULTI_PROPERTY_AD = 'MULTI_PROPERTY_AD',
  UNIT_SHOWCASE_AD = 'UNIT_SHOWCASE_AD',
  SINGLE_PROPERTY_OFFER = 'SINGLE_PROPERTY_OFFER',
  LIMITED_TIME_OFFER = 'LIMITED_TIME_OFFER',
  SEASONAL_OFFER = 'SEASONAL_OFFER',
  MULTI_PROPERTY_OFFERS_GRID = 'MULTI_PROPERTY_OFFERS_GRID',
  OFFERS_CAROUSEL = 'OFFERS_CAROUSEL',
  FLASH_DEALS = 'FLASH_DEALS',
  HORIZONTAL_PROPERTY_LIST = 'HORIZONTAL_PROPERTY_LIST',
  VERTICAL_PROPERTY_GRID = 'VERTICAL_PROPERTY_GRID',
  MIXED_LAYOUT_LIST = 'MIXED_LAYOUT_LIST',
  COMPACT_PROPERTY_LIST = 'COMPACT_PROPERTY_LIST',
  FEATURED_PROPERTIES_SHOWCASE = 'FEATURED_PROPERTIES_SHOWCASE',
  CITY_CARDS_GRID = 'CITY_CARDS_GRID',
  DESTINATION_CAROUSEL = 'DESTINATION_CAROUSEL',
  EXPLORE_CITIES = 'EXPLORE_CITIES',
  PREMIUM_CAROUSEL = 'PREMIUM_CAROUSEL',
  INTERACTIVE_SHOWCASE = 'INTERACTIVE_SHOWCASE',
}

export enum SectionAnimation {
  NONE = 'NONE',
  FADE = 'FADE',
  SLIDE = 'SLIDE',
  SCALE = 'SCALE',
  ROTATE = 'ROTATE',
  PARALLAX = 'PARALLAX',
  SHIMMER = 'SHIMMER',
  PULSE = 'PULSE',
  BOUNCE = 'BOUNCE',
  FLIP = 'FLIP',
}

export enum SectionSize {
  COMPACT = 'COMPACT',
  SMALL = 'SMALL',
  MEDIUM = 'MEDIUM',
  LARGE = 'LARGE',
  EXTRA_LARGE = 'EXTRA_LARGE',
  FULL_SCREEN = 'FULL_SCREEN',
}

export enum SectionTarget {
  PROPERTIES = 'PROPERTIES',
  UNITS = 'UNITS',
}